GBA mode, error management, parameters management
Added a new mode that enable the slot2 and start the NDS, to be able to launch GBA retail cart with the usual NDS menu. Added a clean parameter management to enable every functions. Added a parameter that enable the slot2 with the slot1 when launching NDS rom (usefull for some NDS rom that have bonus/functionnality linked to a GBA cart, like pokemon games)